Start on cleaner style
Started remaking a skeleton of the frontend. This is based of the image Ethan sent in the discord. It isnt finished but this is a start. Eventually I should separate the components into their respective files and fix the css. There also is more work to be done to get the functionality back.